sort item classes into namespaces
