+++ src/main/java/com/lothrazar/cyclicmagic/gui/wand/GuiWandInventory.java
@@ -1,7 +1,7 @@
 package com.lothrazar.cyclicmagic.gui.wand;
 import org.lwjgl.opengl.GL11;
 import com.lothrazar.cyclicmagic.gui.GuiBaseContainer;
-import com.lothrazar.cyclicmagic.item.ItemCyclicWand;
+import com.lothrazar.cyclicmagic.item.tool.ItemCyclicWand;
 import com.lothrazar.cyclicmagic.util.Const;
 import com.lothrazar.cyclicmagic.util.UtilSpellCaster;
 import net.minecraft.client.Minecraft;
