+++ src/main/java/com/lothrazar/cyclicmagic/item/gear/ItemEmeraldSword.java
@@ -1,4 +1,4 @@
-package com.lothrazar.cyclicmagic.item;
+package com.lothrazar.cyclicmagic.item.gear;
 import com.lothrazar.cyclicmagic.IHasRecipe;
 import com.lothrazar.cyclicmagic.module.EmeraldArmorModule;
 import net.minecraft.init.Items;
