+++ src/main/java/com/lothrazar/cyclicmagic/net/PacketMoveBlock.java
@@ -1,5 +1,5 @@
 package com.lothrazar.cyclicmagic.net;
-import com.lothrazar.cyclicmagic.item.ItemToolPush;
+import com.lothrazar.cyclicmagic.item.tool.ItemToolPush;
 import com.lothrazar.cyclicmagic.util.UtilPlaceBlocks;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.entity.player.EntityPlayer;
