+++ src/main/java/com/lothrazar/cyclicmagic/registry/ItemRegistry.java
@@ -7,27 +7,27 @@ import com.lothrazar.cyclicmagic.item.BaseItem;
 import com.lothrazar.cyclicmagic.item.ItemAppleEmerald;
 import com.lothrazar.cyclicmagic.item.ItemChestSack;
 import com.lothrazar.cyclicmagic.item.ItemChestSackEmpty;
-import com.lothrazar.cyclicmagic.item.ItemCyclicWand;
-import com.lothrazar.cyclicmagic.item.ItemEmeraldArmor;
 import com.lothrazar.cyclicmagic.item.ItemEnderBook;
-import com.lothrazar.cyclicmagic.item.ItemToolPearlReuse;
-import com.lothrazar.cyclicmagic.item.ItemToolProspector;
 import com.lothrazar.cyclicmagic.item.ItemFoodCorruptedChorus;
 import com.lothrazar.cyclicmagic.item.ItemFoodCrafting;
 import com.lothrazar.cyclicmagic.item.ItemFoodHeart;
 import com.lothrazar.cyclicmagic.item.ItemFoodInventory;
 import com.lothrazar.cyclicmagic.item.ItemStorageBag;
 import com.lothrazar.cyclicmagic.item.ItemPaperCarbon;
-import com.lothrazar.cyclicmagic.item.ItemSandstoneAxe;
-import com.lothrazar.cyclicmagic.item.ItemSandstonePickaxe;
-import com.lothrazar.cyclicmagic.item.ItemSandstoneSpade;
 import com.lothrazar.cyclicmagic.item.ItemSleepingMat;
 import com.lothrazar.cyclicmagic.item.ItemSproutSeeds;
-import com.lothrazar.cyclicmagic.item.ItemToolHarvest;
-import com.lothrazar.cyclicmagic.item.ItemToolSpawnInspect;
-import com.lothrazar.cyclicmagic.item.ItemToolSpelunker;
-import com.lothrazar.cyclicmagic.item.ItemToolWarp;
+import com.lothrazar.cyclicmagic.item.gear.ItemEmeraldArmor;
+import com.lothrazar.cyclicmagic.item.gear.ItemSandstoneAxe;
+import com.lothrazar.cyclicmagic.item.gear.ItemSandstonePickaxe;
+import com.lothrazar.cyclicmagic.item.gear.ItemSandstoneSpade;
 import com.lothrazar.cyclicmagic.item.projectile.*;
+import com.lothrazar.cyclicmagic.item.tool.ItemCyclicWand;
+import com.lothrazar.cyclicmagic.item.tool.ItemToolHarvest;
+import com.lothrazar.cyclicmagic.item.tool.ItemToolPearlReuse;
+import com.lothrazar.cyclicmagic.item.tool.ItemToolProspector;
+import com.lothrazar.cyclicmagic.item.tool.ItemToolSpawnInspect;
+import com.lothrazar.cyclicmagic.item.tool.ItemToolSpelunker;
+import com.lothrazar.cyclicmagic.item.tool.ItemToolWarp;
 import com.lothrazar.cyclicmagic.util.Const;
 import net.minecraft.item.Item;
 import net.minecraft.util.ResourceLocation;
